Fix naming convention expectation
